Ensures that a call to Do() flushes the command queue
The redigo package ensures that a call to Do() will flush any commands
that were queued via the Sned() method, however a call to Do() on the
mock does not empty the queued commands. This commit ensures that
we recursively empty the queue before handling the command invoked by
Do().