Fix races on Conn fields
In this commit I protect the fields of `Conn` by a lock, to prevent
races.  I went for a single coarse lock around everything, on the theory
that it's easier to get right that way.  I didn't worry too much about
trying not to hold the lock for too long, either: I just had each method
hold the lock, except private methods as commented, and methods that
just call other public methods.  One problem with that is that we do
hold the lock while calling out to user-provided functions (FlushMock,
ResponseHandler, etc.) which we ideally wouldn't do, but it's at least a
start.  Ideally, we might have finer-grained locks for each field, but
that will require more care (and more tests) to get right.

This should fix at least a few existing races; it needs some more tests
(especially around the reply-queue, and Receive/Send) to prove that.