Make unneeded fields no longer exported
This commit is the first in a series to make the package safer for
concurrent use (see #59, #60, #61).  In this commit, I changed a number
of fields which cannot be made safe in their current form to be
unexported.  In particular:
1. the entire `Response` type is now unexported, as it's not used in the
   public API
2. all fields of `Cmd`, and a few more fields of `Conn` are now
   unexported as well
3. the documentation of `Conn` now says that the other fields should not
   be modified after use

For most of those fields in (2), there didn't seem to be any need to
expose that data at all.  The few where there did seem to be a need, I
added getters for.  It would be easy to add getters for others if they
are needed.

An alternative to (3) would be to make those fields private, and add
setters or constructors for them.  That seemed like more ceremony than
it was worth, but it's an option.

I did not make any logical changes; those will come in future commits.