Fix races on command fields
In this commit I protect the fields of `Cmd` by a lock, to prevent
races.  The main one to worry about is with the code now in
`getResponse`, but I tested all the methods.  We also need to
change `Cmd.hash` to a pointer-receiver, so it doesn't try to copy the
whole `Cmd`.  This fixes several races that would previously happen in
`Do()`, as the added tests prove.

I also turned on the race detector for tests.